Add files via upload
This is the c++ code for a Domino game with 2 players, the game allows two players to take turns playing pieces from the available pieces in hand, and trying to match it with the pieces on the table. Any player that would be able to place all of his/her pieces on the table matching the head and tail, will be the winner, or the game will end after a selected number of rounds if none of the players were able to play all of their pieces.